Add command line options and help and refactor, Fix #6
